Assistant checkpoint: Fix cart price storage and display
Assistant generated file changes:
- shared/schema.ts: Update cart item price to be integer
- client/src/components/product-card.tsx: Ensure integer prices
- client/src/pages/admin/carts.tsx: Fix admin cart price display

---

User prompt:

Issue Summary:
Product prices are still not appearing in the admin dashboard cart view.
Prices must be stored along with the submitted cart and should not be missing from the database.
Prices should remain as whole numbers (no decimal conversion) and be displayed as whole numbers.
The total cart price in the admin dashboard should match the pre-submission total from the main page cart.
Expected Behavior:
When a cart is submitted, each product's price must be stored exactly as it was before submission (whole number, no decimal conversion).
The admin dashboard cart should display the same product prices and total price as the main page cart before submission.
Total price calculation in the admin dashboard should be accurate and match the pre-submission cart total.
Debugging & Fixes Needed:
1. Ensure Product Prices Are Stored Correctly in the Database
When a cart is submitted, product prices must be included in the stored cart data.
Ensure that prices are stored as whole numbers and do not get converted to decimals.
Check the database to verify that prices are saved correctly alongside the cart details.
2. Ensure Backend Returns Product Prices for Submitted Carts
The API should return stored product prices instead of re-fetching live product data.
Verify that the API response for the admin dashboard cart includes accurate, whole-number prices for each product.
3. Ensure Admin Dashboard Cart Uses the Same Pricing Data as the Main Page Cart
The admin dashboard cart should retrieve and display prices exactly as they were before submission.
Ensure the frontend is mapping and displaying the correct price field from the API response.
4. Ensure Total Price Matches Pre-Submission Cart Total
The total price in the admin dashboard cart should be calculated from the stored product prices rather than recalculating incorrectly.
Verify that the total price is displayed as a whole number without decimal conversion.

+++ shared/schema.ts
@@ -140,7 +140,7 @@ export const cartItemSchema = z.object({
   description: z.string(),
   images: z.array(z.string()),
   fullImages: z.array(z.string()).optional(),
-  price: z.number().min(0, "Price must be non-negative"),
+  price: z.number().int().min(0, "Price must be non-negative"),
   isAvailable: z.boolean().optional(),
   createdAt: z.string().or(z.date()).optional()
 });
